refactor: 파일트리 조회함수를 ProjectController에서 사용하도록 변경
