+++ src/main/java/goorm/fullstack/webide/service/FolderService.java
@@ -1,7 +1,6 @@
 package goorm.fullstack.webide.service;
 
 import goorm.fullstack.webide.domain.File;
-import goorm.fullstack.webide.dto.FileTreeNodeDto;
 import goorm.fullstack.webide.dto.FolderRenameRequestDto;
 import goorm.fullstack.webide.dto.FolderRequestDto;
 
@@ -12,7 +11,5 @@ public interface FolderService {
     void deleteFolder(int id);
 
     File renameFolder(int id, FolderRenameRequestDto folderRenameRequestDto);
-
-    FileTreeNodeDto getFileTree(int projectId);
 
 }
